inndf:  report buffindexed usage with 2 decimals
Output was previously truncated at the inferior number (e.g. 37.00%).

+++ storage/ovsqlite/ovsqlite.c
@@ -970,7 +970,7 @@ bool ovsqlite_ctl(
     }
     switch (type) {
     case OVSPACE:
-        *(int *)val = -1;
+        *(float *)val = -1.0f;
         return true;
     case OVSORT:
         *(OVSORTTYPE *)val = OVNEWSGROUP;
